# src/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Boilerplate",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080"
        }
    ],
    "paths": {
...
        "/api/v1/auth/magic-link/request": {
            "post": {
                "operationId": "v1.auth.magic.request",
                "tags": [
                    "MagicLink"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Always accepted (anti-enumeration).",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "accepted": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "required": [
                                                "accepted"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmailRequest": {
                "type": "object",
                "description": "Body for forgot-password and magic-link/request: just an email.",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255
                    }
                },
                "required": [
                    "email"
                ],
                "title": "EmailRequest"
            },
...
        },
        "responses": {
...
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}